Fixed the custom datetime parsing
The timezone offset was not correctly applied, because the UTC offset was applied to a local time. Now an UTC date is created and the UTC offset is applied to this date, which seems like the correct behaviour to me.

This method is only used in browsers that don't support the ISO8601 timeformat natively (i.e. Internet Explorer < 9).